Moved tests around
