+++ src/BrowserHost.Tests/Features/Settings/SettingsFeatureTest.cs
